refactor: 각 페이지별 컴포넌트 import 부분 수정
