+++ src/components/basic/PageWrapper/index.jsx
@@ -1,4 +1,4 @@
-import Header from 'components/Header/Header';
+import Header from 'components/Header';
 import Navigation from 'components/Navigation';
 import theme from 'styles/theme';
