--- src/components/basic/index.js
@@ -1,8 +1,0 @@
-export { default as Avatar } from './Avatar';
-export { default as Button } from './Button';
-export { default as Icon } from './Icon';
-export { default as Image } from './Image';
-export { default as Input } from './Input';
-export { default as PageWrapper } from './PageWrapper';
-export { default as Tab } from './Tab';
-export { default as Text } from './Text';
