+++ src/pages/LoginPage/index.jsx
@@ -1,14 +1,10 @@
 import Logo from 'components/Logo';
 import styled from '@emotion/styled';
 import { IMAGE_NAMES } from 'utils/constants/images';
-import LoginForm from 'components/LoginForm';
-import { login } from 'utils/apis/userApi';
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
-import LoginModal from 'components/LoginModal';
 import { useUserContext } from 'contexts/UserContext';
-import PageWrapper from 'components/basic/pageWrapper';
-import Modal from 'components/Modal';
+import { PageWrapper, LoginForm, Modal } from 'components';
 
 const LoginWrapper = styled.div`
   width: 100%;
