+++ src/pages/MainPage/PostBody.jsx
@@ -3,9 +3,7 @@ import { jsx, css } from '@emotion/react';
 import { useCallback, useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from '@emotion/styled';
-import Image from 'components/basic/Image';
-import Text from 'components/basic/Text';
-import Icon from 'components/basic/Icon';
+import { Image, Text, Icon } from 'components';
 import theme from 'styles/theme';
 import { setLike, setDisLike } from 'utils/apis/postApi';
 import { setNotification } from 'utils/apis/userApi';
