+++ src/pages/MainPage/PostHeader.jsx
@@ -2,9 +2,8 @@ import styled from '@emotion/styled';
 import { useState, useCallback, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import IconBtn from './IconButton';
-import Profile from 'components/Profile';
+import { Profile, Modal } from 'components';
 import { useUserContext } from 'contexts/UserContext';
-import Modal from 'components/Modal';
 import theme from 'styles/theme';
 import { deletePost } from 'utils/apis/postApi';
 import useLocalToken from 'hooks/useLocalToken';
