+++ src/pages/MainPage/index.jsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import PageWrapper from 'components/basic/pageWrapper';
+import PageWrapper from 'components/basic/PageWrapper';
 import { getPostsPart } from 'utils/apis/postApi';
 import PostItem from './PostItem';
 import { useState, useEffect, useCallback, useRef } from 'react';
