+++ src/pages/MyInfoPage/UserProfile.jsx
@@ -1,8 +1,6 @@
 import { useEffect, useState, useCallback, useRef } from 'react';
 import styled from '@emotion/styled';
-import Avatar from 'components/basic/Avatar';
-import Text from 'components/basic/Text';
-import Icon from 'components/basic/Icon';
+import { Avatar, Text, Icon } from 'components';
 import { EDIT } from 'utils/constants/icons/names';
 import { IMAGE_URLS } from 'utils/constants/images';
 import ChangeNameForm from 'components/ChangeNameForm';
