+++ src/pages/SearchTagPage/index.jsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
 import { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import PageWrapper from 'components/basic/pageWrapper';
-import TagSearchResult from 'components/TagSearchResult';
+import PageWrapper from 'components/basic/PageWrapper';
+import TagSearchResult from 'pages/SearchPage/TagSearchResult';
 import { searchTag } from 'utils/apis/postApi';
 
 const Header = styled.h3`
