+++ src/pages/UserPage/UserData.jsx
@@ -1,13 +1,9 @@
 import { useEffect, useState, useCallback } from 'react';
-import Text from 'components/basic/Text';
+import { Text, Avatar, Button, Image, Modal } from 'components';
 import { useUserContext } from 'contexts/UserContext';
 import theme from 'styles/theme';
-import Avatar from 'components/basic/Avatar';
-import Button from 'components/basic/Button';
-import Image from 'components/basic/Image';
 import { IMAGE_URLS } from 'utils/constants/images';
 import { useNavigate } from 'react-router-dom';
-import Modal from 'components/Modal';
 import PropTypes from 'prop-types';
 
 import {
